+++ packages/mavis-id-sdk/src/common/id-response.ts
@@ -1,0 +1,4 @@
+export interface IdResponse {
+  id_token: string
+  address?: string
+}